-Tinkered with critter cage
-Saddle overlay for pigs fixed
-More beetroot on a stick variants (added golden variant, pretty much)
-Beetroot on a stick recipes
-Critter cage textures work with dataGen
-Fixed crystal dimension particle bug
-Fixed another crash with the ultimate fishing rod
-Critter cage yet to be fully polished before it being considered "done"

+++ src/main/java/io/github/chaosawakens/common/entity/CarrotPigEntity.java
@@ -86,7 +86,7 @@ public class CarrotPigEntity extends PigEntity implements IAnimatable {
     
     @OnlyIn(Dist.CLIENT)
     public Vector3d getLeashOffset() {
-       return new Vector3d(0.0D, (double)(0.6F * this.getEyeHeight()), (double)(this.getBbWidth() * 0.4F));
+       return new Vector3d(0.0D, (0.6F * this.getEyeHeight()), (this.getBbWidth() * 0.4F));
     }
 
     public boolean isFood(ItemStack stack) {
